requirements.txt
aiogram==2.11.2
attrs==19.1.0
aiograph>=0.2
asyncpg
aioredis==1.3.1
python-dotenv
requests
